+++ src/tests/units/operator_tests.cpp
@@ -22,7 +22,7 @@ void tests(DeviceType device) {
     values.clear();
     std::vector<float> vB({1, 2, 3, 4, 5, 6});
 
-    auto B = graph->param("B", {3, 2}, inits::from_vector(vB));
+    auto B = graph->param("B", {3, 2}, inits::fromVector(vB));
     auto B2 = B * 2.0f;
     graph->forward();
 
@@ -40,8 +40,8 @@ void tests(DeviceType device) {
     std::vector<float> vA({1, -2, 3, -4});
     std::vector<float> vB({0.5, 1.5});
 
-    auto a = graph->constant({2, 2, 1}, inits::from_vector(vA));
-    auto b = graph->constant({2, 1}, inits::from_vector(vB));
+    auto a = graph->constant({2, 2, 1}, inits::fromVector(vA));
+    auto b = graph->constant({2, 1}, inits::fromVector(vB));
 
     auto compare = [&](Expr res, std::function<float(float,float)> f, bool exactMatch) -> bool {
       if (res->shape() != Shape({ 2, 2, 1 }))
@@ -93,7 +93,7 @@ void tests(DeviceType device) {
     std::vector<float> vT4({1, 5, 3, 7, 2, 6, 4, 8});
     std::vector<float> vT5({1, 2, 5, 6, 3, 4, 7, 8});
 
-    auto a = graph->constant({2, 4}, inits::from_vector(vA));
+    auto a = graph->constant({2, 4}, inits::fromVector(vA));
 
     auto t1 = transpose(a);
     auto t2 = transpose(t1);
@@ -144,7 +144,7 @@ void tests(DeviceType device) {
     std::vector<float> lsmOut({ -0.6444f, -0.7444f, -1.10319f, -0.40319f,
                                 -111.45f, 0.0f, -100.05001f, 0.0f });
 
-    auto input = graph->constant({2, 2, 2}, inits::from_vector(in));
+    auto input = graph->constant({2, 2, 2}, inits::fromVector(in));
 
     auto sm  = softmax(input);
     auto lsm = logsoftmax(input);
@@ -183,10 +183,10 @@ void tests(DeviceType device) {
     });
 #endif
 
-    auto a = graph->constant({2, 2, 4}, inits::glorot_uniform);
+    auto a = graph->constant({2, 2, 4}, inits::glorotUniform());
 
-    auto gamma = graph->param("gamma", {1, 4}, inits::ones);
-    auto beta = graph->param("beta", {1, 4}, inits::zeros);
+    auto gamma = graph->param("gamma", {1, 4}, inits::ones());
+    auto beta = graph->param("beta", {1, 4}, inits::zeros());
 
     auto ln = layerNorm(a, gamma, beta);
 
@@ -218,7 +218,7 @@ void tests(DeviceType device) {
     std::vector<float> vL9({8.13364336f, 7.17551536f});  // np.log(np.sum(np.exp(a), axis=1))
     std::vector<float> vW({5.0f, 4.55555556f});          // np.mean(a*s1,axis=-1) / np.mean(s1,axis=-1)
 
-    auto a = graph->constant({2, 4}, inits::from_vector(vA));
+    auto a = graph->constant({2, 4}, inits::fromVector(vA));
 
     auto s1 = sum(a, /*axis=*/ 0);
     auto s2 = sum(a, /*axis=*/ 1);
@@ -292,10 +292,10 @@ void tests(DeviceType device) {
                             3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
                             4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4});
 
-    auto in1 = graph->constant({1, 2, 2, 3}, inits::from_value(1));
-    auto in2 = graph->constant({1, 2, 2, 3}, inits::from_value(2));
-    auto in3 = graph->constant({1, 2, 2, 3}, inits::from_value(3));
-    auto in4 = graph->constant({1, 2, 2, 3}, inits::from_value(4));
+    auto in1 = graph->constant({1, 2, 2, 3}, inits::fromValue(1));
+    auto in2 = graph->constant({1, 2, 2, 3}, inits::fromValue(2));
+    auto in3 = graph->constant({1, 2, 2, 3}, inits::fromValue(3));
+    auto in4 = graph->constant({1, 2, 2, 3}, inits::fromValue(4));
 
     auto c1out1 = concatenate({in1, in2, in3, in4}, /*axis=*/ 2);
     auto c1out2 = concatenate({in1, in2, in3, in4}, /*axis=*/ -1);
@@ -338,8 +338,8 @@ void tests(DeviceType device) {
                            76, 100,
                            103, 136});
 
-    auto A = graph->param("A", {2, 2, 3}, inits::from_vector(vA));
-    auto B = graph->param("B", {3, 2}, inits::from_vector(vB));
+    auto A = graph->param("A", {2, 2, 3}, inits::fromVector(vA));
+    auto B = graph->param("B", {3, 2}, inits::fromVector(vB));
     auto C = dot(A, B);
 
     CHECK(C->shape() == Shape({2, 2, 2}));
@@ -361,9 +361,9 @@ void tests(DeviceType device) {
                             4, 5, 6, 2.3, 6.7,
                             7, 8, 9, 3.4, 7.8,
                             1, 1, 2, 4.5, 8.9});
-      auto S  = graph->param("S",  { 2, 4 }, inits::from_vector(vS));
-      auto D  = graph->param("D",  { 4, 5 }, inits::from_vector(vD));
-      auto DT = graph->param("DT", { 5, 4 }, inits::from_vector(vD)); // example matrix with transposed dimensions
+      auto S  = graph->param("S",  { 2, 4 }, inits::fromVector(vS));
+      auto D  = graph->param("D",  { 4, 5 }, inits::fromVector(vD));
+      auto DT = graph->param("DT", { 5, 4 }, inits::fromVector(vD)); // example matrix with transposed dimensions
       std::vector<float> SV;    // create CSR version of S
       std::vector<IndexType> SI, SO;
       SO.push_back((IndexType)SI.size());
@@ -382,15 +382,15 @@ void tests(DeviceType device) {
       auto STxSxDd = dot(S, SxDd, /*transA=*/true);
       auto SxDs = csr_dot( // sparse x dense
             S->shape(),
-            graph->constant({(int)SV.size()}, inits::from_vector(SV), Type::float32),
-            graph->constant({(int)SI.size()}, inits::from_vector(SI), Type::uint32),
-            graph->constant({(int)SO.size()}, inits::from_vector(SO), Type::uint32),
+            graph->constant({(int)SV.size()}, inits::fromVector(SV), Type::float32),
+            graph->constant({(int)SI.size()}, inits::fromVector(SI), Type::uint32),
+            graph->constant({(int)SO.size()}, inits::fromVector(SO), Type::uint32),
             D);
       auto STxSxDs = csr_dot(   // transpose(sparse) x dense; we use result of previous since dimensions match
             S->shape(),
-            graph->constant({(int)SV.size()}, inits::from_vector(SV), Type::float32),
-            graph->constant({(int)SI.size()}, inits::from_vector(SI), Type::uint32),
-            graph->constant({(int)SO.size()}, inits::from_vector(SO), Type::uint32),
+            graph->constant({(int)SV.size()}, inits::fromVector(SV), Type::float32),
+            graph->constant({(int)SI.size()}, inits::fromVector(SI), Type::uint32),
+            graph->constant({(int)SO.size()}, inits::fromVector(SO), Type::uint32),
             SxDd, /*transS=*/true);
 
       auto DTxSTd   = dot(DT,     S, /*transA=*/false, /*transB=*/true);
@@ -398,16 +398,16 @@ void tests(DeviceType device) {
       auto DTxSTs = dot_csr( // dense x sparse
             DT,
             S->shape(),
-            graph->constant({(int)SV.size()}, inits::from_vector(SV), Type::float32),
-            graph->constant({(int)SI.size()}, inits::from_vector(SI), Type::uint32),
-            graph->constant({(int)SO.size()}, inits::from_vector(SO), Type::uint32),
+            graph->constant({(int)SV.size()}, inits::fromVector(SV), Type::float32),
+            graph->constant({(int)SI.size()}, inits::fromVector(SI), Type::uint32),
+            graph->constant({(int)SO.size()}, inits::fromVector(SO), Type::uint32),
             /*transS=*/true);
       auto DTxSTxSs = dot_csr( // dense x transpose(sparse)
             DTxSTd,
             S->shape(),
-            graph->constant({(int)SV.size()}, inits::from_vector(SV), Type::float32),
-            graph->constant({(int)SI.size()}, inits::from_vector(SI), Type::uint32),
-            graph->constant({(int)SO.size()}, inits::from_vector(SO), Type::uint32));
+            graph->constant({(int)SV.size()}, inits::fromVector(SV), Type::float32),
+            graph->constant({(int)SI.size()}, inits::fromVector(SI), Type::uint32),
+            graph->constant({(int)SO.size()}, inits::fromVector(SO), Type::uint32));
 
       CHECK(SxDs->shape() == SxDd->shape());
       CHECK(STxSxDs->shape() == STxSxDd->shape());
@@ -432,9 +432,9 @@ void tests(DeviceType device) {
     std::vector<float> vB({1, 2, 3, 4, 5, 6});
     std::vector<float> vAff({24, 30, 51, 66, 78, 102, 105, 138});
 
-    auto A = graph->param("A", {4, 3}, inits::from_vector(vA));
-    auto B = graph->param("B", {3, 2}, inits::from_vector(vB));
-    auto C = graph->param("C", {4, 2}, inits::from_value(2));
+    auto A = graph->param("A", {4, 3}, inits::fromVector(vA));
+    auto B = graph->param("B", {3, 2}, inits::fromVector(vB));
+    auto C = graph->param("C", {4, 2}, inits::fromValue(2));
     auto aff1 = affine(A, B, C);
     auto aff2 = dot(A, B) + C;
     graph->forward();
@@ -457,7 +457,7 @@ void tests(DeviceType device) {
     std::vector<float> vB({1, 2, 3, 4, 5, 6, 1, 2, 3, 4, 5, 6});
     std::vector<float> vC({1, 2, 3, 1, 2, 3, 4, 5, 6, 4, 5, 6});
 
-    auto A = graph->param("A", {2,3}, inits::from_vector(vA));
+    auto A = graph->param("A", {2,3}, inits::fromVector(vA));
     auto I = repeat(A, 1, 0);
     auto B = repeat(A, 2, 0);
     auto C = repeat(A, 2, 1);
@@ -482,9 +482,9 @@ void tests(DeviceType device) {
 
     std::vector<float> vIn({1, 2, 3, 4, 5, 6, 7, 8});
 
-    auto A = graph->param("A", {2, 4}, inits::from_vector(vIn));
+    auto A = graph->param("A", {2, 4}, inits::fromVector(vIn));
     auto Af = flatten(A);
-    auto B = graph->param("B", {2, 2, 1, 2}, inits::from_vector(vIn));
+    auto B = graph->param("B", {2, 2, 1, 2}, inits::fromVector(vIn));
     auto Bf = flatten(B);
     graph->forward();
 
@@ -517,7 +517,7 @@ void tests(DeviceType device) {
     std::vector<float> vB4({4, 5, 6, 4, 5, 6});
     std::vector<float> vB6;
 
-    auto A = graph->param("A", {4, 3}, inits::from_vector(vA));
+    auto A = graph->param("A", {4, 3}, inits::fromVector(vA));
     auto B0 = rows(A, iB0);
     auto B1 = rows(A, iB1);
     auto B2 = rows(A, iB2);
@@ -577,7 +577,7 @@ void tests(DeviceType device) {
     std::vector<float> vB4({2, 2, 6, 6, 10, 10});
     std::vector<float> vB6;
 
-    auto A = graph->param("A", {3, 4}, inits::from_vector(vA));
+    auto A = graph->param("A", {3, 4}, inits::fromVector(vA));
     auto B0 = cols(A, iB0);
     auto B1 = cols(A, iB1);
     auto B2 = cols(A, iB2);
@@ -624,10 +624,10 @@ void tests(DeviceType device) {
     std::vector<float> vA({0, .3333, -.2, -.3, 0, 4.5, 5.2, -10, 101.45, -100.05, 0, 1.05e-5});
     std::vector<IndexType> idx({0, 1});
 
-    auto A1 = graph->param("4x3", {4,3}, inits::from_vector(vA));
+    auto A1 = graph->param("4x3", {4,3}, inits::fromVector(vA));
     auto B1 = rows(transpose(A1), idx);
     auto C1 = transpose(cols(A1, idx));
-    auto A2 = graph->param("6x2", {6,2}, inits::from_vector(vA));
+    auto A2 = graph->param("6x2", {6,2}, inits::fromVector(vA));
     auto B2 = cols(transpose(A2), idx);
     auto C2 = transpose(rows(A2, idx));
     graph->forward();
@@ -675,7 +675,7 @@ void tests(DeviceType device) {
     std::vector<float> vS2({-4, 5, -6, 7, -8, 9});
     std::vector<float> vS3({7, -8, 9, -10, 11, -12});
 
-    auto A = graph->param("4x3", {4,3}, inits::from_vector(vA));
+    auto A = graph->param("4x3", {4,3}, inits::fromVector(vA));
     auto B1a = index_select(A, 0, IndexVector({0})); // always uses gather()
     auto B1b = slice(A,  0, 0);                        // memory-consecutive view
     auto B2  = slice(A,  1, 0);                        // not memory-consecutive
@@ -690,7 +690,7 @@ void tests(DeviceType device) {
     CHECK(B4b->type() == "sliceView"); // must use view
     CHECK(B5.get() == A.get());        // must be no-op
 
-    auto C = graph->param("2x3x2", {2, 3, 2}, inits::from_vector(vC));
+    auto C = graph->param("2x3x2", {2, 3, 2}, inits::fromVector(vC));
     auto D1 = slice(C,  0, 0);
     auto D2 = slice(C, -2, 2);
     auto D3 = index_select(C, 1, IndexVector({0, 2})); // C[:,(0,2),:]
@@ -698,7 +698,7 @@ void tests(DeviceType device) {
     CHECK(D2->type() == "gather");
     // enable this once gather() supports batched indices:
     //auto D4 = gather(C, 1, graph->constant({2, 2, 1}, // [C[0,(2,1),:],C[1,(0,2),:]]
-    //                                       inits::from_vector(std::vector<IndexType>{
+    //                                       inits::fromVector(std::vector<IndexType>{
     //                                         2, 1,
     //                                         0, 2 }),
     //                                       Type::uint32));
@@ -734,7 +734,7 @@ void tests(DeviceType device) {
     std::vector<float> vA({0, .3333, -.2, -.3, 0, 4.5, 5.2, -10, 101.45, -100.05, 0, 1.05e-5});
     std::vector<IndexType> indices({0, 2});
 
-    auto A = graph->param("4x3", {4, 3}, inits::from_vector(vA));
+    auto A = graph->param("4x3", {4, 3}, inits::fromVector(vA));
     auto B1 = rows(A, indices);
     auto B2 = gather(A, 0, graph->indices(indices, A, 0));
     auto C1 = cols(A, indices);